[PORT FROM JBMR0] AsfExtractor: MediaScan can't work again after adb push some speical file to phone
BZ: 72174

from my test it happens in the AsfExtractor.cpp.
It will in initialize() => for() forever.
Because the objectSize =0
then offset will be == fix value,
then readAt() == 24,
so forever.
In common, offset + objectSize will reach another header.
But this bad file(may be made by somebody) is different, it reachs a bad another header.
So it can happen this issue.
If you want to know the bad file name,you can refer patch:(80715) and (79686 Patch Set 9).

Org-Change-Id: I55cdb5be94ad28e92fcfb64594e1e5d13d6c633e
Signed-off-by: Kun Jiang <kunx.jiang@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>

Change-Id: I8401ce21fc8cbf98bf741006ce5142c89ce35434
Reviewed-on: http://android.intel.com:8080/83782
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Feng, Wei <wei.feng@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>