Remove a workaround in VPP which may cause deadlock.[MERR-VV]
BZ: 97149

When locking and unlocking DUT during video playback, AwesomePlayer
will restart and seek to previous time point to resume. With VPP enabled,
there is a flag mFirstFrameoDone to indicate the first VPP output is
ready, and if this flag is not set, VPPProcessor will not send out frames
to render. However, in AwesomePlayer, if no output available, mSeek flag
will not be cleared, which will cause dead lock in seek logic.

Change-Id: I96b8640dc4d49049910fba6d511cef157cf57053
Reviewed-on: http://android.intel.com:8080/99684
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yu, Linda <linda.yu@intel.com>
Reviewed-by: Wang, Lili A <lili.a.wang@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>