Fix a green thumbnail problem for one WMV clip
BZ: 101341 109551

For thumbnail generation, the ASF extractor should give a key frame
to HW decoder. But for the given WMV clip, the ASF extractor gives
a P-frame to MediaSDK. MediaSDK can't handle such a situation, thus
a garbage thumbnail is generated.

In this patch, asf extractor will wait until a key-frame is receved.

Change-Id: I3f1bc60e3d9859184d006ffd33415b497113bcaf
Signed-off-by: wfeng6 <wei.feng@intel.com>
Reviewed-on: http://android.intel.com:8080/107834
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Chen, Tianmi <tianmi.chen@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>