VPP: fix a race condition issue when VPP threads exit
BZ: 113263

mFillThread->mRunCond.signal();
mFillThread->requestExitAndWait();

The upper sequence has pontential dead lock issue. The thread that is requested to exit
may be waiting on the condition "mRunCond" in threadLoop() and the meanwhile the main
thread that invoke requestExitAndWait() may be waiting on the condition "mThreadExitedCondition".
So we must invoke requestExit() first to set mExitPending true before signal the condition
"mRunCond" to avoid the race condition.

Change-Id: Iad11651909f6424e042fc14ed6e635d57ad22c1c
Signed-off-by: Jason Hu <jason.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/111826
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>