[Movie studio]: fix the messy thumbnail when imported one special clip(1)
BZ: 113431

It's android defect.
root cause:
When MP4 container doesn't have stss table it assumes all samples are
sync samples. So it's possible to seek to a p frame and see messy thumbnail.
Use the meta data type kKeyFakeSync to check whether MP4 container has stss
table or not.
this patch is for CTP/Merrifield platform

Change-Id: I02652dcb33b73b828021bbcaeb9da418b927dbe9
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-on: http://android.intel.com:8080/113762
Reviewed-by: Feng, Wei <wei.feng@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>